refactor: file structure
This commit efactors the file structure to organize files better.
