+++ src/pages/mod.rs
@@ -1,0 +1,3 @@
+pub mod overview;
+pub mod processes;
+pub mod resources;
